Refactor tree rewriter
